requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astroid==2.15.6
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==4.1.0
bleach==6.1.0
blinker==1.6.2
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.34.0
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.3
ctransformers==0.2.27
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.6
django-autocomplete-light==3.9.7
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-environ==0.10.0
django-import-export==3.2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.3
Flask==3.0.0
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gunicorn==21.2.0
huggingface-hub==0.17.3
idna==3.4
inflection==0.5.1
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
kombu==5.3.2
lazy-object-proxy==1.9.0
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
nltk==3.8.1
notebook==7.0.5
notebook_shim==0.2.3
numpy==1.24.1
odfpy==1.4.1
openai==0.28.1
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.18
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2023.9.1
platformdirs==3.10.0
postgis==1.0.4
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.3.1
scikit-surprise==1.1.3
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tablib==3.5.0
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.14.1
tomlkit==0.12.1
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
torchvision==0.16.0+cpu
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.1
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
virtualenv==20.24.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.3.7
whitenoise==6.5.0
widgetsnbextension==4.0.9
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
